# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SkyMP Backend API
  version: 2.0.0
  description: Canonical API v2 for launchers, dashboards, and SkyMP game servers.
servers:
  - url: http://localhost:4000
...
paths:
...
  /api/v2/admin/players/{profileId}:
    parameters:
      - $ref: '#/components/parameters/ProfileId'
    get:
      tags: [Administration]
      operationId: getPlayer
      security: [{ dashboardBearer: [] }]
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
    put:
      tags: [Administration]
      operationId: updatePlayer
      security: [{ dashboardBearer: [] }]
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
...
  /api/v2/admin/content/{resource}:
    parameters:
      - name: resource
        in: path
        required: true
        description: lore, rules, or whitelist-notes
        schema: { type: string, enum: [lore, rules, whitelist-notes] }
    get:
      tags: [Administration]
      operationId: getManagedContent
      security: [{ dashboardBearer: [] }]
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
    put:
      tags: [Administration]
      operationId: updateManagedContent
      security: [{ dashboardBearer: [] }]
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
...
  /api/v2/game-servers/{key}/profiles/{id}/factions:
    post:
      tags: [Game servers]
      operationId: assignProfileFaction
      parameters:
        - { $ref: '#/components/parameters/ServerKey' }
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { $ref: '#/components/parameters/AuthTokenHeader' }
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '201': { $ref: '#/components/responses/JsonSuccess' }
  /api/v2/game-servers/{key}/profiles/{id}/factions/{assignmentId}:
    delete:
      tags: [Game servers]
      operationId: removeProfileFaction
      parameters:
        - { $ref: '#/components/parameters/ServerKey' }
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: assignmentId, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/AuthTokenHeader' }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
...
components:
  securitySchemes:
    dashboardBearer:
      type: http
      scheme: bearer
  parameters:
    ServerKey:
      name: key
      in: path
      required: true
      schema: { type: string }
    ProfileId:
      name: profileId
      in: path
      required: true
      schema: { type: integer }
...
    AuthTokenHeader:
      name: X-Auth-Token
      in: header
      required: true
      schema: { type: string }
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  responses:
    JsonSuccess:
      description: Successful JSON response
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
...